+++ cli/abeamer-cli.ts
@@ -166,6 +166,16 @@ namespace Cli {
       `arguments to be passed to ffmpeg after the arguments passed by abeamer`,
   };
 
+  argOpts['gifPre'] = {
+    param: 'array', allowOption: true, desc:
+      `arguments to be passed to gif generator(convert) before the arguments passed by abeamer`,
+  };
+
+  argOpts['gifPost'] = {
+    param: 'array', allowOption: true, desc:
+      `arguments to be passed to gif generator(convert) after the arguments passed by abeamer`,
+  };
+
   // ------------------------------------------------------------------------
   //                               Print Usage
   // ------------------------------------------------------------------------
@@ -557,12 +567,20 @@ To modify the fps, edit the [js/main.ts] file.
 
     const cmdLine = 'convert';
 
-    const args = ['-delay', `1x${report.fps}`];
+    let args = ['-delay', `1x${report.fps}`];
     const loop = argOpts['loop'].value as string || '0';
     args.push('-loop', loop);
     if (toOptimize) { args.push('-strip', '-layers', 'optimize', '-alpha', 'deactivate'); }
 
     args.push(report.framespattern.replace(/\%\d*d/, '*'), gifFileName);
+
+    if (argOpts['gifPre'].multipleValue) {
+      args = [...argOpts['gifPre'].multipleValue as string[], ...args];
+    }
+
+    if (argOpts['gifPost'].multipleValue) {
+      args = [...args, ...argOpts['gifPost'].multipleValue as string[]];
+    }
 
     if (isVerbose) {
       console.log(`\n${cmdLine} ${args.join(' ')}\n`);