+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add support value parameters starting with -- in the abeamer cli.
 - Add `--movie-pre` and `--movie-post` parameters to abeamer cli.
   > These arguments to be passed to ffmpeg before/after the arguments passed by abeamer
+- Add `--gif-pre` and `--gif-post` parameters to abeamer cli.
+  > These arguments to be passed to convert before/after the arguments passed by abeamer
 
 ## [1.0.1] -
 ### Fix